requirements.txt
alembic==1.14.0
fastapi==0.115.6
markdown2==2.5.2
passlib==1.7.4
pydantic==2.10.5
python-dotenv==1.0.1
python_jose==3.3.0
SQLAlchemy==2.0.36
uvicorn==0.34.0
